[chore] 헥사고날 아키텍처 적용
- 기존 api 모듈 -> infra 모듈로 통합
- infra 모듈 : 외부 시스템 연동 역할 수행
