+++ http/movie.http
@@ -1,2 +1,2 @@
 ###
-GET http://localhost:8080/api/movie
+GET http://localhost:8080/api/movies
