--- movie-api/src/main/resources/application-local.yml
@@ -1,14 +1,0 @@
-spring:
-  config:
-    activate:
-      on-profile: local
-
-  datasource:
-    url: jdbc:mysql://localhost:3306/movie-local
-    username: movie
-    password: movie
-    driver-class-name: com.mysql.jdbc.Driver
-  jpa:
-    hibernate:
-      ddl-auto: create-drop
-    show-sql: true
