+++ movie-business/src/main/kotlin/com/example/business/movie/repository/MovieRepository.kt
@@ -5,6 +5,6 @@ import java.time.LocalDate
 
 interface MovieRepository {
 
-    fun findMoviesReleasedUntil(now: LocalDate): List<Movie>
+    fun getMoviesReleasedUntil(now: LocalDate): List<Movie>
 
 }
