+++ movie-common/build.gradle.kts
@@ -1,0 +1,3 @@
+dependencies {
+    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
+}
