+++ movie-infrastructure/build.gradle.kts
@@ -1,4 +1,5 @@
 dependencies {
+    implementation(project(":movie-application"))
     implementation(project(":movie-business"))
 
     implementation("org.springframework.boot:spring-boot-starter-web")
