--- movie-infrastructure/src/main/kotlin/com/example/infrastructure/common/config/JpaConfig.kt
@@ -1,9 +1,0 @@
-package com.example.infrastructure.common.config
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing
-
-@Configuration
-@EnableJpaAuditing
-class JpaConfig {
-}
