--- movie-infrastructure/src/main/kotlin/com/example/infrastructure/movie/jpa/MovieJpaRepository.kt
@@ -1,10 +1,0 @@
-package com.example.infrastructure.movie.jpa
-
-import org.springframework.data.jpa.repository.JpaRepository
-import java.time.LocalDate
-
-interface MovieJpaRepository: JpaRepository<MovieEntity, Long> {
-
-    fun findByReleaseDateLessThanEqualOrderByReleaseDateAsc(now: LocalDate): List<MovieEntity>
-
-}
