--- movie-infrastructure/src/main/kotlin/com/example/infrastructure/reservation/jpa/ReservationStatus.kt
@@ -1,5 +1,0 @@
-package com.example.infrastructure.reservation.jpa
-
-enum class ReservationStatus {
-    DONE
-}
