--- movie-infrastructure/src/main/kotlin/com/example/infrastructure/seat/jpa/SeatStatus.kt
@@ -1,6 +1,0 @@
-package com.example.infrastructure.seat.jpa
-
-enum class SeatStatus {
-    AVAILABLE,
-    RESERVED
-}
