--- movie-infrastructure/src/main/kotlin/com/example/infrastructure/theater/jpa/TheaterEntity.kt
@@ -1,18 +1,0 @@
-package com.example.infrastructure.theater.jpa
-
-import com.example.infrastructure.common.BaseEntity
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.GenerationType
-import jakarta.persistence.Id
-
-@Entity
-class TheaterEntity (
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "theater_id")
-    val id: Long = 0,
-
-    val name: String
-): BaseEntity()
