--- movie-infrastructure/src/main/kotlin/com/example/infrastructure/theater/jpa/TheaterJpaRepository.kt
@@ -1,9 +1,0 @@
-package com.example.infrastructure.theater.jpa
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface TheaterJpaRepository: JpaRepository<TheaterEntity, Long> {
-
-    fun findByIdIn(theaterIds: Set<Long>): List<TheaterEntity>
-
-}
