--- movie-infrastructure/src/main/kotlin/com/example/infrastructure/user/jpa/UserEntity.kt
@@ -1,17 +1,0 @@
-package com.example.infrastructure.user.jpa
-
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.GenerationType
-import jakarta.persistence.Id
-
-@Entity
-class UserEntity (
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
-    val id: Long = 0,
-
-    val name: String
-)
