+++ settings.gradle.kts
@@ -1,6 +1,6 @@
 rootProject.name = "movie-multi-module"
 
-include("movie-api")
+include("movie-common")
 include("movie-business")
 include("movie-infrastructure")
 include("movie-application")
